# packages/bc-wallet-openapi/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Credential Showcase API
  version: 0.1.1
servers:
  - url: https://virtserver.swaggerhub.com/SphereonInt/showcase/0.1.0/v1
    description: SwaggerHub API Auto Mocking
  - url: https://bcshowcase-api-dev.nborbit.ca/
...
security:
  - OAuth2: [ ]
  - OIDC: [ ]
paths:
  /scenarios/issuances:
    get:
      tags:
        - Issuance Scenario
      summary: List all issuance scenarios
      operationId: listIssuanceScenarios
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuanceScenariosResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Issuance Scenario
      summary: Create a new issuance scenario
      operationId: createIssuanceScenario
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssuanceScenarioRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuanceScenarioResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /scenarios/issuances/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [ "Issuance Scenario" ]
      summary: Retrieve a single issuance scenario
      operationId: getIssuanceScenario
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuanceScenarioResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [ "Issuance Scenario" ]
      summary: Update an issuance scenario (overwrite)
      operationId: updateIssuanceScenario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssuanceScenarioRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuanceScenarioResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [ "Issuance Scenario" ]
      summary: Delete an issuance scenario
      operationId: deleteIssuanceScenario
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /scenarios/issuances/{slug}/steps:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - "Issuance Scenario"
      summary: List steps in an issuance scenario
      operationId: listIssuanceScenarioSteps
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [ "Issuance Scenario" ]
      summary: Add a step to an issuance scenario
      operationId: createIssuanceScenarioStep
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StepRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Step'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /scenarios/issuances/{slug}/steps/{stepId}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
      - name: stepId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [ "Issuance Scenario" ]
      summary: Retrieve a specific step in an issuance scenario
      operationId: getIssuanceScenarioStep
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Step'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [ "Issuance Scenario" ]
      summary: Overwrite a step in an issuance scenario
      operationId: updateIssuanceScenarioStep
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Step'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Step'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [ "Issuance Scenario" ]
      summary: Delete a step in an issuance scenario
      operationId: deleteIssuanceScenarioStep
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /scenarios/issuances/{slug}/steps/{stepId}/actions:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
      - name: stepId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [ "Issuance Scenario" ]
      summary: List actions in a step
      operationId: listStepActions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepActionsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [ "Issuance Scenario" ]
      summary: Create an action in a step
      operationId: createStepAction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StepActionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepAction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /scenarios/issuances/{slug}/steps/{stepId}/actions/{actionId}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
      - name: stepId
        in: path
        required: true
        schema:
          type: string
      - name: actionId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [ "Issuance Scenario" ]
      summary: Retrieve a single action
      operationId: getStepAction
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepAction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [ "Issuance Scenario" ]
      summary: Overwrite an action
      operationId: updateStepAction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StepActionRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepAction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [ "Issuance Scenario" ]
      summary: Delete an action
      operationId: deleteStepAction
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: <https://auth.example.com/oauth/authorize>
          tokenUrl: <https://auth.example.com/oauth/token>
          scopes:
            editor: Grants editor access
            admin: Grants admin access
            viewer: Grants viewer access
    OIDC:
      type: openIdConnect
      openIdConnectUrl: <https://auth.example.com/.well-known/openid-configuration>
      description: OpenID Connect security scheme for OAuth2 flows
  schemas:
...
    Asset:
      type: object
      required:
        - id
        - mediaType
        - content
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Unique identifier for the asset in the database
          example: 123e4567-e89b-12d3-a456-426614174469
        mediaType:
          type: string
          description: Type of asset (e.g., image, document, icon)
          example: image/jpeg
        content:
          type: string
          description: Binary data of the asset
          example: base64 encoded binary data
        fileName:
          type: string
          description: Name of the asset file
          example: asset.jpg
        description:
          type: string
          description: Description of the asset
          example: A beautiful image of a cat
        createdAt:
          type: string
          description: Date and time the asset was created
          format: date-time
          example: 2025-03-06T11:30:16.573Z
        updatedAt:
          type: string
          description: Date and time the asset was last updated
          format: date-time
          example: 2025-03-06T11:30:16.573Z
...
    Step:
      type: object
      required:
        - id
        - title
        - description
        - order
        - type
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Unique identifier for the step
          example: 123e4567-e89b-12d3-a456-434314174000
        title:
          type: string
          description: Title of the step
          example: Verify Identity
        description:
          type: string
          description: Detailed description of the step
          example: Verify the user's identity
        order:
          type: integer
          description: Order of the step in the scenario
          minimum: 0
          example: 1
        type:
          $ref: '#/components/schemas/StepType'
        subScenario:
          type: string
          description: Optional sub-scenario for this step
          example: 123e4567-e89b-12d3-a456-434314174000
        actions:
          type: array
          items:
            $ref: '#/components/schemas/StepAction'
        asset:
          $ref: '#/components/schemas/Asset'
        createdAt:
          type: string
          description: Date and time the step was created
          format: date-time
          example: 2025-03-06T11:30:16.573Z
        updatedAt:
          type: string
          description: Date and time the step was last updated
          format: date-time
          example: 2025-03-06T11:30:16.573Z
    StepRequest:
      type: object
      required:
        - title
        - description
        - order
        - type
        - actions
      properties:
        title:
          type: string
          description: Title of the step
          example: Verify Identity
        description:
          type: string
          description: Detailed description of the step
          example: Verify the user's identity
        order:
          type: integer
          description: Order of the step in the scenario
          minimum: 0
          example: 1
        type:
          $ref: '#/components/schemas/StepType'
        subScenario:
          type: string
          description: Optional sub-scenario for this step
          example: 123e4567-e89b-12d3-a456-434314174000
        actions:
          type: array
          description: List of actions associated with this step
          items:
            oneOf:
              - $ref: '#/components/schemas/StepActionRequest'
        asset:
          type: string
          description: Asset referenced by this step
          example: 456e4567-e89b-12d3-a456-426614174000
    StepsResponse:
      properties:
        steps:
          type: array
          items:
            $ref: '#/components/schemas/Step'
...
    StepType:
      type: string
      enum:
        - HUMAN_TASK
        - SERVICE
        - SCENARIO
      description: Type of step in the scenario
      example: HUMAN_TASK
    StepAction:
      type: object
      required:
        - id
        - actionType
        - title
        - text
      #        - createdAt  FIXME This gives issues with the individual step types, they do not all have *Request variants yet
      #        - updatedAt
      properties:
        id:
          type: string
          description: Unique identifier for the action
          example: 123e4567-ef2d-12d3-abcd-426614174456
        actionType:
          $ref: '#/components/schemas/StepActionType'
        title:
          type: string
          description: Title of the action
          example: Connect Wallet
        text:
          type: string
          description: Descriptive text for the action
          example: Connect your wallet to continue
        createdAt:
          type: string
          description: Date and time the step action was created
          format: date-time
          example: 2025-03-06T11:30:16.573Z
        updatedAt:
          type: string
          description: Date and time the step action was last updated
          format: date-time
          example: 2025-03-06T11:30:16.573Z
      discriminator:
        propertyName: actionType
        mapping:
          ARIES_OOB: '#/components/schemas/AriesOOBAction'
          ACCEPT_CREDENTIAL: '#/components/schemas/AcceptCredentialAction'
          SETUP_CONNECTION: '#/components/schemas/SetupConnectionAction'
          CHOOSE_WALLET: '#/components/schemas/ChooseWalletAction'
          BUTTON: '#/components/schemas/ButtonAction'
    StepActionRequest:
      type: object
      required:
        - actionType
        - title
        - text
      properties:
        title:
          type: string
          description: Title of the action
          example: Connect Wallet
        text:
          type: string
          description: Descriptive text for the action
          example: Connect your wallet to continue
        actionType:
          $ref: '#/components/schemas/StepActionType'
      discriminator:
        propertyName: actionType
        mapping:
          ARIES_OOB: '#/components/schemas/AriesOOBActionRequest'
          ACCEPT_CREDENTIAL: '#/components/schemas/AcceptCredentialActionRequest'
          SETUP_CONNECTION: '#/components/schemas/SetupConnectionActionRequest'
          CHOOSE_WALLET: '#/components/schemas/ChooseWalletActionRequest'
          BUTTON: '#/components/schemas/ButtonActionRequest'
    StepActionsResponse:
      properties:
        actions:
          type: array
          items:
            $ref: '#/components/schemas/StepAction'
...
    StepActionType:
      type: string
      enum:
        - ARIES_OOB
        - ACCEPT_CREDENTIAL
        - SHARE_CREDENTIAL
        - SETUP_CONNECTION
        - CHOOSE_WALLET
        - BUTTON
      description: Type of action
      example: ARIES_OOB
...
    AriesOOBActionRequest:
      allOf:
        - $ref: '#/components/schemas/StepActionRequest'
        - type: object
          required:
            - proofRequest
            - credentialDefinitionId
          properties:
            proofRequest:
              $ref: '#/components/schemas/AriesProofRequestRequest'
            credentialDefinitionId:
              type: string
              description: Credential definition ID
              example: 123e4567-e89b-12d3-a456-426614174456
...
    ScenarioRequest:
      type: object
      required:
        - name
        - type
        - description
        - steps
        - personas
      properties:
        name:
          type: string
          description: Name of the scenario
          example: Credential Issuance
        description:
          type: string
          description: Detailed description of the scenario
          example: This scenario issues credentials to users
        steps:
          type: array
          description: Ordered list of steps in the scenario
          items:
            $ref: '#/components/schemas/StepRequest'
        personas:
          type: array
          description: References to personas used in this scenario
          items:
            type: string
            example: 123e4567-e89b-12d3-a456-426614174456
        hidden:
          type: boolean
          description: Whether the workflow is hidden from the user
          example: false
    IssuanceScenariosResponse:
      type: object
      properties:
        issuanceScenarios:
          type: array
          items:
            $ref: '#/components/schemas/IssuanceScenario'
    IssuanceScenarioResponse:
      type: object
      properties:
        issuanceScenario:
          $ref: '#/components/schemas/IssuanceScenario'
    IssuanceScenario:
      type: object
      allOf:
        - $ref: '#/components/schemas/Scenario'
        - type: object
          #         required:
          #           - issuer # This cannot be here because of the allOf to ScenarioRequest. It's a inheritance issue
          properties:
            issuer:
              $ref: '#/components/schemas/Issuer'
            type:
              type: string
              enum:
                - "ISSUANCE"
      description: Scenario specialization for credential issuance
    IssuanceScenarioRequest:
      type: object
      allOf:
        - $ref: '#/components/schemas/ScenarioRequest'
        - type: object
          #          required:
          #            - issuer # This cannot be here because of the allOf to ScenarioRequest. It's a inheritance issue
          properties:
            issuer:
              type: string
              example: "313e4567-e89b-12d3-a456-426614174469"
      description: Scenario specialization for credential issuance
...
  responses:
    'BadRequest':
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Invalid request data'
    'Unauthorized':
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Invalid authentication credentials'
    'Forbidden':
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Insufficient permissions to access resource'
    'NotFound':
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Resource not found'
    'InternalServerError':
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'An unexpected error occurred'